+++ src/Life/Main/Add.hs
@@ -13,7 +13,7 @@ import Path.IO (copyDirRecur, copyFile, doesDirExist, doesFileExist, ensureDir, 
 
 import Life.Configuration (LifeConfiguration, LifePath (..), directories, files, parseHomeLife,
                            writeGlobalLife)
-import Life.Github (Branch(..), addToRepo, withSynced)
+import Life.Github (addToRepo, master, withSynced)
 import Life.Main.Init (lifeInitQuestion)
 import Life.Message (abortCmd, errorMessage, infoMessage, warningMessage)
 import Life.Shell (LifeExistence (..), relativeToHome, repoName, whatIsLife)
@@ -25,7 +25,7 @@ import qualified Data.Set as Set
 lifeAdd :: LifePath -> IO ()
 lifeAdd lPath = whatIsLife >>= \case
     -- actual life add process
-    Both _ _ -> withSynced (Branch "master") addingProcess
+    Both _ _ -> withSynced master addingProcess
 
     -- if one of them is missing -- abort
     OnlyRepo _ -> abortCmd "add" ".life file doesn't exist"