+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The change log is available [on GitHub][2].
 
 # 1.0.2
 
-* Branch as newtype.
+* [#42](https://github.com/kowainik/life-sync/issues/42):
+  Create newtype `Branch`.
 
 # 1.0.1